Each step runs only one animation frame and any queued tasks
- If processing a step generates tasks that wait for
an animation frame, (only) these are executed at the
end of the step. This prevents endless loops in the
case of a program that subscribes to onAnimationFrame 

We had to do this because we cannot tell when a program
update returns Cmd.none since it could be wrapped with
Cmd.map. So we cannot tell when a program update cycle
is complete.

So we must handle asynchronous steps explicitly
- Add FileReader proxy that can cancel step timer
- Cancel step timer when processing a blob download

And this led to discovering the possibility of
endless looping so must be careful when running
animation frame tasks during a step.